Move hyperdrive, yieldSources, token, chains, out of appconfig/ and into own feature folders
