--- apps/hyperdrive-trading/src/appconfig/chains/supportedChains.ts
@@ -1,3 +1,0 @@
-export const supportedChainIds = [42069, 31337] as const;
-
-export type SupportedChainId = typeof supportedChainIds[number];
