+++ apps/hyperdrive-trading/src/chains/supportedChains.ts
@@ -1,0 +1,3 @@
+export const supportedChainIds = [42069, 31337] as const;
+
+export type SupportedChainId = (typeof supportedChainIds)[number];
