+++ apps/hyperdrive-trading/src/hyperdrive/formatHyperdriveName.ts
