+++ apps/hyperdrive-trading/src/token/TokenConfig.ts
