+++ apps/hyperdrive-trading/src/ui/appconfig/useAppConfig.ts
@@ -2,8 +2,8 @@ import { useQuery } from "@tanstack/react-query";
 import assertNever from "assert-never";
 import { LocalAddressesJson } from "src/addresses/LocalAddressesJson";
 import { AppConfig } from "src/appconfig/AppConfig";
-import { SupportedChainId } from "src/appconfig/chains/supportedChains";
 import { getAppConfigFromLocalAddresses } from "src/appconfig/getAppConfigFromLocalAddresses";
+import { SupportedChainId } from "src/chains/supportedChains";
 import { PublicClient } from "viem";
 import { useChainId, usePublicClient } from "wagmi";
