+++ apps/hyperdrive-trading/src/ui/hyperdrive/hooks/useCurrentFixedAPR.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { HyperdriveConfig } from "src/appconfig/HyperdriveConfig";
+import { HyperdriveConfig } from "src/hyperdrive/HyperdriveConfig";
 
 import { formatRate } from "src/base/formatRate";
 import { makeQueryKey } from "src/base/makeQueryKey";
