+++ apps/hyperdrive-trading/src/ui/hyperdrive/hooks/useLpApy.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import { HyperdriveConfig } from "src/appconfig/HyperdriveConfig";
 import { makeQueryKey } from "src/base/makeQueryKey";
+import { HyperdriveConfig } from "src/hyperdrive/HyperdriveConfig";
 import { usePoolConfig } from "src/ui/hyperdrive/hooks/usePoolConfig";
 import { usePoolInfo } from "src/ui/hyperdrive/hooks/usePoolInfo";
 import { useReadHyperdrive } from "src/ui/hyperdrive/hooks/useReadHyperdrive";
