+++ apps/hyperdrive-trading/src/ui/hyperdrive/longs/hooks/useCurrentLongPrice.ts
@@ -1,6 +1,6 @@
 import { QueryStatus, useQuery } from "@tanstack/react-query";
-import { HyperdriveConfig } from "src/appconfig/HyperdriveConfig";
 import { makeQueryKey } from "src/base/makeQueryKey";
+import { HyperdriveConfig } from "src/hyperdrive/HyperdriveConfig";
 import { useReadHyperdrive } from "src/ui/hyperdrive/hooks/useReadHyperdrive";
 import { formatUnits } from "viem";
