+++ apps/hyperdrive-trading/src/ui/hyperdrive/lp/hooks/usePreviewRemoveLiquidity.ts
@@ -1,6 +1,6 @@
 import { MutationStatus, useQuery } from "@tanstack/react-query";
-import { HyperdriveConfig } from "src/appconfig/HyperdriveConfig";
 import { makeQueryKey } from "src/base/makeQueryKey";
+import { HyperdriveConfig } from "src/hyperdrive/HyperdriveConfig";
 import { useReadWriteHyperdrive } from "src/ui/hyperdrive/hooks/useReadWriteHyperdrive";
 import { Address } from "wagmi";
