+++ apps/hyperdrive-trading/src/ui/hyperdrive/shorts/hooks/useMaxShort.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import * as dnum from "dnum";
-import { HyperdriveConfig } from "src/appconfig/HyperdriveConfig";
 import { makeQueryKey } from "src/base/makeQueryKey";
+import { HyperdriveConfig } from "src/hyperdrive/HyperdriveConfig";
 import { useAppConfig } from "src/ui/appconfig/useAppConfig";
 import { useReadHyperdrive } from "src/ui/hyperdrive/hooks/useReadHyperdrive";
