+++ apps/hyperdrive-trading/src/ui/markets/MarketDetailsBody/MarketHeader.tsx
@@ -1,5 +1,5 @@
 import { ReactElement } from "react";
-import { HyperdriveConfig } from "src/appconfig/HyperdriveConfig";
+import { HyperdriveConfig } from "src/hyperdrive/HyperdriveConfig";
 import { PriceBadges } from "src/ui/markets/MarketDetailsBody/PriceBadges";
 
 export function MarketHeader({
