+++ apps/hyperdrive-trading/src/ui/markets/MarketDetailsBody/PriceBadges.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from "react";
-import { HyperdriveConfig } from "src/appconfig/HyperdriveConfig";
 import { divideBigInt } from "src/base/divideBigInt";
 import { parseUnits } from "src/base/parseUnits";
+import { HyperdriveConfig } from "src/hyperdrive/HyperdriveConfig";
 import { formatBalance } from "src/ui/base/formatting/formatBalance";
 
 export function PriceBadges({
